:sparkles: :alembic: Projects
Hardcoded a number of things due to the very specific nature of the projects.
In simple terms, this code allocates days off for the project if applicable and schedules project meetings on these days. This only applies to the bachelor program, since the master projects are not scheduled.
This code can partially be used for making dummy courses in the future.
Note 1: The user is expected to indicate whether or not these project constraints should be used AND the user is expected to indicate which block it is (>>>GUI?)

Note 2: I had to add a default value to  the room booking at the bottom, default value is now -1 to suppress errors.

Future: Will implement this constraint in a similar way in other scheduling algorithms later today.